Change: remove sendNotification from all randomAccessDataProviders
Rather than using sendNotification, we add problems to initializationResult
and throw where appropriate within individual methods.

Rather than assuming a global notification flow, these non-ui classes expose
their error and allow the callers to handle it for display. This also allows
the callers to clear the error if it is no longer relevant.